Split apart handlers code into separate runInTerminal.ts and runOnMarkdown.ts files. Continued refactor.
